refactor: move `offset` into option
